# api-docs.yml
...
openapi: 3.0.3
info:
  title: UltraZend SMTP Server API
  description: |
    # 🚀 UltraZend SMTP Server API
    
    **Servidor SMTP profissional com API REST completa para envio e gerenciamento de emails.**
    
    ## 🌟 Funcionalidades
    
    - **Envio de emails** via API REST
    - **Servidor SMTP** completo (MX + Submission)  
    - **Autenticação JWT** com refresh tokens
    - **Rate limiting** inteligente
    - **Webhooks** para notificações
    - **Analytics** em tempo real
    - **Sistema de filas** Redis Bull
    - **Monitoramento** Prometheus
    - **Segurança** multicamadas
    
    ## 🔐 Autenticação
    
    A API utiliza **JWT Bearer tokens** para autenticação:
    
    ```bash
    curl -H "Authorization: Bearer YOUR_JWT_TOKEN" \
         -H "Content-Type: application/json" \
         https://api.ultrazend.com.br/api/emails/send
    ```
    
    ## 🚦 Rate Limiting
    
    - **API**: 1000 requests/hora por usuário
    - **Email Send**: 10,000 emails/hora por usuário
    - **Auth**: 10 tentativas/15min por IP
    
    ## 📧 SMTP Configuration
    
    **MX Server (recebimento):**
    - Host: `mail.ultrazend.com.br`
    - Port: `25`
    
    **Submission Server (envio autenticado):**
    - Host: `mail.ultrazend.com.br`
    - Port: `587`
    - Security: `STARTTLS`
    
    ## 🔗 Links Úteis
    
    - **Dashboard**: https://www.ultrazend.com.br
    - **Status Page**: https://status.ultrazend.com.br
    - **Support**: suporte@ultrazend.com.br
...
    
  version: 2.0.0
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtido através do endpoint de login.
        
        Exemplo: `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
    
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API Key para integração direta sem login.
        
        Exemplo: `X-API-Key: ultrazend_ak_1234567890abcdef`
...
paths:
...
  /auth/verify-email:
    post:
      tags:
        - Authentication
      summary: Verificar email
      description: |
        Verifica o email do usuário usando o token enviado por email.
        
        Após a verificação, o usuário pode fazer login normalmente.
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  minLength: 64
                  maxLength: 64
                  example: "abc123def456789012345678901234567890123456789012345678901234"
                  description: "Token de 64 caracteres recebido por email"
              required:
                - token
      responses:
        '200':
          description: Email verificado com sucesso
        '400':
          description: Token inválido ou expirado

  # === ENVIO DE EMAILS ===
...